Crate rustc_public_bridge

Crate that implements what will become the rustc side of rustc_public.

This crate serves as a proxy for making calls to rustc queries.

This crate is not intended to be invoked directly by users. This crate is the public API of rustc that will be invoked by the rustc_public crate.

For more information see https://github.com/rust-lang/project-stable-mir

Note

This API is still completely unstable and subject to change.

Modules

alloc

Internal memory allocator implementation for rustc_public.

bridge

Defines a set of traits that is used for abstracting rustc_public’s components that are needed in rustc_public_bridge.

builder 🔒

Logic required to produce a monomorphic body.

context

Implementation of CompilerCtxt.

rustc_internalDeprecated

Structs

Container

A container which is used for TLS.

IndexMap

Similar to rustc’s FxIndexMap, IndexMap with extra safety features added.

Tables

Traits

Bridge

A trait defining types that are used to emulate rustc_public components, which is really useful when programming in rustc_public-agnostic settings.

IndexedVal

Functions

filter_def_ids 🔒

Iterate over the definitions of the given crate.