rustc_public_bridge

Module alloc

Internal memory allocator implementation for rustc_public.

This module handles all direct interactions with rustc queries and performs the actual memory allocations. The stable interface in rustc_public::alloc delegates all query-related operations to this implementation.

Functions

allocation_filter

Creates an Allocation only from information within the AllocRange.

create_ty_and_layout

try_new_indirect

try_new_scalar

try_new_slice